refactor: optimize `_transformAmount` logic in Money class and improve associated test structure
